+++ ompi/mca/coll/tuned/coll_tuned_decision_fixed.c
@@ -2,7 +2,7 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2009 The University of Tennessee and The University
+ * Copyright (c) 2004-2012 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart, 
@@ -56,7 +56,7 @@ ompi_coll_tuned_allreduce_intra_dec_fixed (void *sbuf, void *rbuf, int count,
      * Ring algorithm does not support non-commutative operations.
      */
     ompi_datatype_type_size(dtype, &dsize);
-    block_dsize = dsize * count;
+    block_dsize = dsize * (ptrdiff_t)count;
 
     if (block_dsize < intermediate_message) {
         return (ompi_coll_tuned_allreduce_intra_recursivedoubling (sbuf, rbuf, 
@@ -66,7 +66,7 @@ ompi_coll_tuned_allreduce_intra_dec_fixed (void *sbuf, void *rbuf, int count,
 
     if( ompi_op_is_commute(op) && (count > comm_size) ) {
         const size_t segment_size = 1 << 20; /* 1 MB */
-        if ((comm_size * segment_size >= block_dsize)) {
+        if (((size_t)comm_size * (size_t)segment_size >= block_dsize)) {
             return (ompi_coll_tuned_allreduce_intra_ring (sbuf, rbuf, count, dtype, 
                                                           op, comm, module));
         } else {
@@ -116,7 +116,7 @@ int ompi_coll_tuned_alltoall_intra_dec_fixed(void *sbuf, int scount,
        Has better performance for messages of intermediate sizes than the old one */
     /* determine block size */
     ompi_datatype_type_size(sdtype, &dsize);
-    block_dsize = dsize * scount;
+    block_dsize = dsize * (ptrdiff_t)scount;
 
     if ((block_dsize < 200) && (communicator_size > 12)) {
         return ompi_coll_tuned_alltoall_intra_bruck(sbuf, scount, sdtype, 
@@ -360,7 +360,7 @@ int ompi_coll_tuned_reduce_intra_dec_fixed( void *sendbuf, void *recvbuf,
 
     /* need data size for decision function */
     ompi_datatype_type_size(datatype, &dsize);
-    message_size = dsize * count;   /* needed for decision */
+    message_size = dsize * (ptrdiff_t)count;   /* needed for decision */
 
     /**
      * If the operation is non commutative we currently have choice of linear 
@@ -535,7 +535,7 @@ int ompi_coll_tuned_allgather_intra_dec_fixed(void *sbuf, int scount,
 
     /* Determine complete data size */
     ompi_datatype_type_size(sdtype, &dsize);
-    total_dsize = dsize * scount * communicator_size;   
+    total_dsize = dsize * (ptrdiff_t)scount * (ptrdiff_t)communicator_size;   
    
     OPAL_OUTPUT((ompi_coll_tuned_stream, "ompi_coll_tuned_allgather_intra_dec_fixed"
                  " rank %d com_size %d msg_length %lu",
@@ -632,7 +632,7 @@ int ompi_coll_tuned_allgatherv_intra_dec_fixed(void *sbuf, int scount,
     ompi_datatype_type_size(sdtype, &dsize);
     total_dsize = 0;
     for (i = 0; i < communicator_size; i++) {
-        total_dsize += dsize * rcounts[i];
+        total_dsize += dsize * (ptrdiff_t)rcounts[i];
     }
     
     OPAL_OUTPUT((ompi_coll_tuned_stream, 
@@ -697,10 +697,10 @@ int ompi_coll_tuned_gather_intra_dec_fixed(void *sbuf, int scount,
     /* Determine block size */
     if (rank == root) {
         ompi_datatype_type_size(rdtype, &dsize);
-        block_size = dsize * rcount;
+        block_size = dsize * (ptrdiff_t)rcount;
     } else {
         ompi_datatype_type_size(sdtype, &dsize);
-        block_size = dsize * scount;
+        block_size = dsize * (ptrdiff_t)scount;
     }
 
     if (block_size > large_block_size) {
@@ -758,10 +758,10 @@ int ompi_coll_tuned_scatter_intra_dec_fixed(void *sbuf, int scount,
     /* Determine block size */
     if (root == rank) {
         ompi_datatype_type_size(sdtype, &dsize);
-        block_size = dsize * scount;
+        block_size = dsize * (ptrdiff_t)scount;
     } else {
         ompi_datatype_type_size(rdtype, &dsize);
-        block_size = dsize * rcount;
+        block_size = dsize * (ptrdiff_t)rcount;
     } 
 
     if ((communicator_size > small_comm_size) &&
